requirements.txt
allure-pytest==2.8.13
allure-python-commons==2.8.13
chardet==3.0.4
docutils==0.15.2
execnet==1.7.1
jmespath==0.10.0
more-itertools==8.5.0
namedlist==1.8
packaging==20.4
parse==1.17.0
parse-type==0.5.2
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==5.4.1
pytest-base-url==1.4.1
pytest-bdd==3.2.1
pytest-html==1.22.0
pytest-localserver==0.5.0
pytest-metadata==1.10.0
pytest-testdox==1.2.1
pytest-variables==1.7.1
pytest-xdist==1.29.0
python-dateutil==2.8.1
python-social-auth==0.3.6
python3-openid==3.2.0
requests==2.22.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
termcolor==1.1.0
urllib3==1.25.10
wcwidth==0.2.5
zipp==3.1.0